requirements.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.10.0
Pillow>=8.3.1
clip @ git+https://github.com/openai/CLIP.git
transformers>=4.20.0